firefox issue appears to only occur in development mode. will create pull request!
