+++ src/modules/FilteredLearningModule.js
@@ -1,24 +1,21 @@
-import React from 'react';
-
-import ModuleMetadata from '../data/ModuleMetadata';
-import LearningModuleListItem from './LearningModuleListItem';
+import React from "react";
+import { CssBaseline } from "@mui/material";
+import ModuleMetadata from "../data/ModuleMetadata";
+import LearningModuleListItem from "./LearningModuleListItem";
 
 function FilteredLearningModule(props) {
-    
-    const learningModuleListItemComponents = ModuleMetadata
-        .filter(metaDatum => metaDatum.difficulty === props.filterKey)
-        .map(metaDatum =>
-            <LearningModuleListItem
-                key={metaDatum.id}
-                moduleMetadata={metaDatum}
-            />
-        )
+  const learningModuleListItemComponents = ModuleMetadata.filter(
+    (metaDatum) => metaDatum.difficulty === props.filterKey
+  ).map((metaDatum) => (
+    <LearningModuleListItem key={metaDatum.id} moduleMetadata={metaDatum} />
+  ));
 
-    return (
-        <>
-            {learningModuleListItemComponents}
-        </>
-    )
+  return (
+    <>
+      <CssBaseline />
+      {learningModuleListItemComponents}
+    </>
+  );
 }
 
 export default FilteredLearningModule;
