+++ src/modules/LearningModuleListItem.js
@@ -1,5 +1,6 @@
 import React from "react";
 import { Link } from "react-router-dom";
+import { CssBaseline } from "@mui/material";
 import Card from "@mui/material/Card";
 import Typography from "@mui/material/Typography";
 import Container from "@mui/material/Container";
@@ -17,37 +18,45 @@ function LearningModuleListItem(props) {
   });
 
   return (
-    <CacheProvider value={cache}>
-      <Card raised className="module-list-item">
-        <Link to={"/main-content/learning-module/module-id=" + props.moduleMetadata.id}>
-          <Container className="module-list-item-inner-container">
-            <Typography
-              variant="h2"
-              sx={{ fontFamily: "Roboto", fontWeight: "500" }}
-            >
-              {props.moduleMetadata.title}
-            </Typography>
-            <Typography variant="h5" sx={{ fontFamily: "Roboto" }}>
-              {props.moduleMetadata.description}
-            </Typography>
-            <Container className="keyword-container">
-              {props.moduleMetadata.keyWords.map((keyword, index) => {
-                return (
-                  <Button
-                    size="small"
-                    variant="contained"
-                    className="keyword-chip"
-                    key={index}
-                  >
-                    {keyword}
-                  </Button>
-                );
-              })}
+    <>
+      <CssBaseline />
+      <CacheProvider value={cache}>
+        <Card raised className="module-list-item">
+          <Link
+            to={
+              "/main-content/learning-module/module-id=" +
+              props.moduleMetadata.id
+            }
+          >
+            <Container className="module-list-item-inner-container">
+              <Typography
+                variant="h2"
+                sx={{ fontFamily: "Roboto", fontWeight: "500" }}
+              >
+                {props.moduleMetadata.title}
+              </Typography>
+              <Typography variant="h5" sx={{ fontFamily: "Roboto" }}>
+                {props.moduleMetadata.description}
+              </Typography>
+              <Container className="keyword-container">
+                {props.moduleMetadata.keyWords.map((keyword, index) => {
+                  return (
+                    <Button
+                      size="small"
+                      variant="contained"
+                      className="keyword-chip"
+                      key={index}
+                    >
+                      {keyword}
+                    </Button>
+                  );
+                })}
+              </Container>
             </Container>
-          </Container>
-        </Link>
-      </Card>
-    </CacheProvider>
+          </Link>
+        </Card>
+      </CacheProvider>
+    </>
   );
 }
